Break system.py into multiple files
